requirements.txt
alembic==1.5.4
apturl==0.5.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dominate==2.3.1
entrypoints==0.3
Faker==6.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.2.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Moment==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.2
greenlet==1.0.0
httplib2==0.14.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
pbr==5.5.1
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.17.0
python-editor==1.0.4
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
SQLAlchemy==1.3.23
systemd-python==234
testresources==2.0.1
text-unidecode==1.3
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
visitor==0.1.3
wadllib==1.3.3
Werkzeug==0.16.1
WTForms==2.1
xkit==0.0.0
